# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Food & Health Planner",
  "description": "Privater Essensplan und Einkaufsliste für 2 Personen.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/components": {
   "get": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Komponenten: Liste (mit Varianten und Nährwerten)",
    "operationId": "list_components_api_components_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ComponentOut"
         },
         "type": "array",
         "title": "Response List Components Api Components Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Komponente anlegen",
    "operationId": "create_component_api_components_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ComponentIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ComponentOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/components/seed-defaults": {
   "post": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Standard-Komponenten für das Frühstück anlegen (idempotent)",
    "description": "Legt zu den im Zutaten-Katalog gefundenen Zutaten die Standard-Komponenten an. Nicht gefundene Namen stehen in `missing`. Vorhandene Komponenten gleichen Namens werden übersprungen.",
    "operationId": "seed_defaults_api_components_seed_defaults_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SeedOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/components/meal": {
   "post": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Baukasten-Mahlzeit berechnen (Nährwerte, Sättigung, Passung ins Tagesziel)",
    "description": "Mengen je Komponente als `grams` oder `units` (Stück × Gewicht je Einheit der Variante). Mit `slot` (und optional `date`) kommt die Passung ins Slot-Ziel der angemeldeten Person dazu, ohne Portionsskalierung.",
    "operationId": "compute_meal_api_components_meal_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MealIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MealOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/components/{component_id}": {
   "get": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Komponente lesen",
    "operationId": "get_component_api_components__component_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "component_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Component Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ComponentOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Komponente ändern",
    "operationId": "update_component_api_components__component_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "component_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Component Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ComponentPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ComponentOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Komponente löschen (409, falls noch verwendet)",
    "operationId": "delete_component_api_components__component_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "component_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Component Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/components/{component_id}/variants": {
   "post": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Variante anlegen",
    "operationId": "create_variant_api_components__component_id__variants_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "component_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Component Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VariantIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VariantOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/components/{component_id}/variants/{variant_id}": {
   "patch": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Variante ändern",
    "operationId": "update_variant_api_components__component_id__variants__variant_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "component_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Component Id"
      }
     },
     {
      "name": "variant_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Variant Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VariantPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VariantOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Komponenten (Frühstücks-Baukasten)"
    ],
    "summary": "Variante löschen",
    "operationId": "delete_variant_api_components__component_id__variants__variant_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "component_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Component Id"
      }
     },
     {
      "name": "variant_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Variant Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ComponentIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Name"
     },
     "kind": {
      "type": "string",
      "enum": [
       "protein",
       "carb",
       "dairy",
       "fruit",
       "veg",
       "other"
      ],
      "title": "Kind"
     },
     "ingredient_id": {
      "type": "integer",
      "exclusiveMinimum": 0.0,
      "title": "Ingredient Id"
     },
     "min_g": {
      "type": "number",
      "maximum": 10000.0,
      "minimum": 0.0,
      "title": "Min G",
      "default": 0.0
     },
     "max_g": {
      "type": "number",
      "maximum": 10000.0,
      "exclusiveMinimum": 0.0,
      "title": "Max G",
      "default": 500.0
     },
     "step_g": {
      "type": "number",
      "maximum": 10000.0,
      "exclusiveMinimum": 0.0,
      "title": "Step G",
      "default": 10.0
     },
     "typical_g": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Typical G"
     },
     "weekend_fixed": {
      "type": "boolean",
      "title": "Weekend Fixed",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "name",
     "kind",
     "ingredient_id"
    ],
    "title": "ComponentIn"
   },
   "ComponentOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "ingredient_id": {
      "type": "integer",
      "title": "Ingredient Id"
     },
     "ingredient_name": {
      "type": "string",
      "title": "Ingredient Name"
     },
     "min_g": {
      "type": "number",
      "title": "Min G"
     },
     "max_g": {
      "type": "number",
      "title": "Max G"
     },
     "step_g": {
      "type": "number",
      "title": "Step G"
     },
     "typical_g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Typical G"
     },
     "weekend_fixed": {
      "type": "boolean",
      "title": "Weekend Fixed"
     },
     "per100": {
      "$ref": "#/components/schemas/Per100Out"
     },
     "variants": {
      "items": {
       "$ref": "#/components/schemas/VariantOut"
      },
      "type": "array",
      "title": "Variants"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "kind",
     "ingredient_id",
     "ingredient_name",
     "min_g",
     "max_g",
     "step_g",
     "typical_g",
     "weekend_fixed",
     "per100",
     "variants"
    ],
    "title": "ComponentOut"
   },
   "ComponentPatch": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "kind": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "protein",
         "carb",
         "dairy",
         "fruit",
         "veg",
         "other"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Kind"
     },
     "ingredient_id": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Ingredient Id"
     },
     "min_g": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Min G"
     },
     "max_g": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max G"
     },
     "step_g": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Step G"
     },
     "typical_g": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Typical G"
     },
     "weekend_fixed": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weekend Fixed"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "ComponentPatch",
    "description": "Nur gesendete Felder werden geändert; `typical_g: null` löscht die übliche Menge."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MealIn": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/MealItemIn"
      },
      "type": "array",
      "maxItems": 50,
      "minItems": 1,
      "title": "Items"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date",
      "description": "Stichtag für die Passung, Standard: heute"
     },
     "slot": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "breakfast",
         "lunch",
         "dinner",
         "snack"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Slot",
      "description": "Slot für die Passung ins Tagesziel"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "items"
    ],
    "title": "MealIn"
   },
...
   "MealOut": {
    "properties": {
     "total": {
      "$ref": "#/components/schemas/NutrientsOut"
     },
     "weight_g": {
      "type": "number",
      "title": "Weight G"
     },
     "energy_density_kcal_per_100g": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Energy Density Kcal Per 100G"
     },
     "satiety": {
      "$ref": "#/components/schemas/SatietyOut"
     },
     "coverage": {
      "type": "number",
      "title": "Coverage"
     },
     "missing": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Missing"
     },
     "lines": {
      "items": {
       "$ref": "#/components/schemas/MealLineOut"
      },
      "type": "array",
      "title": "Lines"
     },
     "notes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notes"
     },
     "fit": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/MealFitOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "total",
     "weight_g",
     "energy_density_kcal_per_100g",
     "satiety",
     "coverage",
     "missing",
     "lines",
     "notes",
     "fit"
    ],
    "title": "MealOut"
   },
...
   "SatietyOut": {
    "properties": {
     "score": {
      "type": "number",
      "title": "Score"
     },
     "parts": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Parts"
     }
    },
    "type": "object",
    "required": [
     "score",
     "parts"
    ],
    "title": "SatietyOut"
   },
   "SeedOut": {
    "properties": {
     "created": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Created"
     },
     "skipped": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Skipped"
     },
     "missing": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Missing"
     },
     "variants_created": {
      "type": "integer",
      "title": "Variants Created"
     },
     "missing_variants": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Missing Variants"
     },
     "matched": {
      "additionalProperties": {
       "type": "string"
      },
      "type": "object",
      "title": "Matched"
     }
    },
    "type": "object",
    "required": [
     "created",
     "skipped",
     "missing",
     "variants_created",
     "missing_variants",
     "matched"
    ],
    "title": "SeedOut"
   },
...
   "VariantIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 120,
      "minLength": 1,
      "title": "Name"
     },
     "ingredient_id": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Ingredient Id",
      "description": "Standard: Zutat der Komponente"
     },
     "grams_per_unit": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Grams Per Unit",
      "description": "z. B. 1 Ei = 60 g"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "name"
    ],
    "title": "VariantIn"
   },
   "VariantOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "component_id": {
      "type": "integer",
      "title": "Component Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "ingredient_id": {
      "type": "integer",
      "title": "Ingredient Id"
     },
     "ingredient_name": {
      "type": "string",
      "title": "Ingredient Name"
     },
     "grams_per_unit": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grams Per Unit"
     },
     "per100": {
      "$ref": "#/components/schemas/Per100Out"
     }
    },
    "type": "object",
    "required": [
     "id",
     "component_id",
     "name",
     "ingredient_id",
     "ingredient_name",
     "grams_per_unit",
     "per100"
    ],
    "title": "VariantOut"
   },
   "VariantPatch": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "ingredient_id": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Ingredient Id"
     },
     "grams_per_unit": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 10000.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Grams Per Unit"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "VariantPatch"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "Zugangstoken der Person (`python -m app.cli`)",
    "scheme": "bearer"
   }
  }
 }
}